pyproject.toml
[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[project]
name = "jwtoxide"
requires-python = ">=3.9"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]

[tool.maturin]
features = ["pyo3/extension-module"]
strip = true

[project.optional-dependencies]
test = ["pytest", "pyjwt>=2.0.0"]
changelog = ["towncrier"]
docs = ["sphinx", "autodoc", "sphinx-rtd-theme"]
lint = ["ruff"]
version = ["bump-my-version"]

[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = '\nversion = "{current_version}"'
replace = '\nversion = "{new_version}"'
regex = true
ignore_missing_version = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "Cargo.toml"
